FILE: moves DifficultyOption and ExpansionOption to option.ts
