+++ libs/spirit-islander/util/src/lib/game/expansions.ts
@@ -1,8 +1,3 @@
 import { EXPANSIONS } from './data';
-import type { Option } from './option';
 
 export type ExpansionName = (typeof EXPANSIONS)[number];
-
-export interface ExpansionOption<TName extends string> extends Option<TName> {
-  expansion?: ExpansionName;
-}
