+++ libs/spirit-islander/util/src/lib/game/maps.ts
@@ -1,6 +1,5 @@
 import { MAPS } from './data';
-import type { DifficultyOption } from './difficulty';
-import type { ExpansionOption } from './expansions';
+import { DifficultyOption, ExpansionOption } from './option';
 
 export interface Map
   extends DifficultyOption<MapName>,
